# shared/openapi.yaml
openapi: 3.1.0
info:
  title: NOMAD Centinela API
  version: 0.1.0
  description: |
    API estable para web y Flutter. Fase 0: datos desde Supabase;
    endpoints de agente devuelven `mock: true` hasta Fase 1.
servers:
  - url: http://localhost:3001
    description: Backend local
paths:
  /api/health:
    get:
      summary: Health check
      operationId: health
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  supabase:
                    type: boolean
                  minimax:
                    type: boolean
                  make_webhook:
                    type: boolean
                  make_api:
                    type: boolean
                  mock:
                    type: boolean
  /api/webhooks/make/ingest:
    post:
      summary: Ingesta de señal desde Make.com (sin PII)
      operationId: makeIngest
      description: |
        Requiere header `X-Nomad-Webhook-Secret` o `Authorization: Bearer <MAKE_WEBHOOK_SECRET>`.
        Crea `exposure_events` con `source_type=make_webhook` y `status=pending_review`.
      security:
        - MakeWebhookSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MakeIngestRequest'
      responses:
        '201':
          description: Evento creado
        '200':
          description: Duplicado (mismo external_id)
        '400':
          description: Validación o PII rechazada
        '401':
          description: Secret inválido
        '503':
          description: MAKE_WEBHOOK_SECRET no configurado
  /api/webhooks/make/health:
    get:
      summary: Health del webhook Make (misma autenticación)
      operationId: makeWebhookHealth
      security:
        - MakeWebhookSecret: []
      responses:
        '200':
          description: Configurado y autorizado
        '401':
          description: Secret inválido
        '503':
          description: No configurado
  /api/institutions:
    get:
      summary: Lista instituciones monitoreadas
      operationId: listInstitutions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Institution'
                  mock:
                    type: boolean
  /api/events:
    get:
      summary: Lista eventos de exposición
      operationId: listEvents
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending_review, approved, rejected, published]
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExposureEvent'
                  mock:
                    type: boolean
  /api/events/{id}:
    get:
      summary: Detalle de evento con trazas
      operationId: getEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '404':
          description: Not found
  /api/citizen/check:
    post:
      summary: Verificar exposición por prefijo de hash (k-anonymity)
      operationId: citizenCheck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hash_prefix]
              properties:
                hash_prefix:
                  type: string
                  minLength: 5
                  maxLength: 5
                  description: Primeros 5 caracteres SHA-1 del email en minúsculas
      responses:
        '200':
          description: Resultado (sin PII)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitizenCheckResult'
  /api/playbooks/{slug}:
    get:
      summary: Obtener playbook de remediación
      operationId: getPlaybook
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          description: Not found
  /api/playbooks/search:
    get:
      summary: Búsqueda FTS de playbooks (RAG Fase 1)
      operationId: searchPlaybooks
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
            maximum: 20
      responses:
        '200':
          description: Resultados FTS (vacío posible)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Playbook'
                  mock:
                    type: boolean
                  query:
                    type: string
        '400':
          description: Falta parámetro q
  /api/agent/chat:
    post:
      summary: Chat ciudadano. Streaming SSE solo en local; en Vercel siempre JSON.
      operationId: agentChat
      parameters:
        - name: stream
          in: query
          required: false
          schema:
            type: string
            enum: ['true', 'false']
          description: |
            `stream=true` activa SSE (UIMessage stream del AI SDK).
            En producción (Vercel) se ignora y siempre devuelve JSON.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                messages:
                  type: array
                  description: |
                    Acepta dos formatos: ChatMessage `{role, content}` o `UIMessage` del AI SDK.
                  items:
                    type: object
      responses:
        '200':
          description: |
            JSON `{role, content, mock, provider?, model?}` cuando no hay stream;
            o `text/event-stream` (UIMessage) si stream=true en local.
        '502':
          description: Falla MiniMax (solo cuando key configurada)
  /api/agent/triage:
    post:
      summary: Clasifica una señal o evento y actualiza severidad (Triage agent)
      operationId: agentTriage
      description: |
        Acepta `event_id` (lee el evento de Supabase y lo actualiza en modo real)
        o `signal` (input manual). En modo mock no muta el evento.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_id:
                  type: string
                  format: uuid
                signal:
                  type: object
                  properties:
                    title:
                      type: string
                    summary:
                      type: string
                    institution_slug:
                      type: string
                    source_type:
                      type: string
                      enum: [osint_feed, hibp, public_report, make_webhook]
      responses:
        '200':
          description: Triage OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriageResponse'
        '404':
          description: event_id no encontrado
        '502':
          description: Falla del modelo
  /api/agent/investigate:
    post:
      summary: Investigator agent (OSINT interno + comparación + recomendación HITL)
      operationId: agentInvestigate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_id]
              properties:
                event_id:
                  type: string
                  format: uuid
                triage:
                  type: object
                  description: Output previo de Triage (opcional)
      responses:
        '200':
          description: Investigación + recomendación HITL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigatorResponse'
        '404':
          description: event_id no encontrado
        '502':
          description: Falla del modelo
  /api/agent/pipeline:
    post:
      summary: Triage + Investigator encadenados (helper para demo)
      operationId: agentPipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_id:
                  type: string
                  format: uuid
                signal:
                  type: object
      responses:
        '200':
          description: Combinado triage + investigación
        '400':
          description: Provide event_id or signal
        '404':
          description: event_id no encontrado
components:
  securitySchemes:
    MakeWebhookSecret:
      type: apiKey
      in: header
      name: X-Nomad-Webhook-Secret
  schemas:
    MakeIngestRequest:
      type: object
      required: [institution_slug, title]
      properties:
        institution_slug:
          type: string
          description: Slug en tabla institutions (ej. digecam)
        title:
          type: string
        summary:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        credentials_count:
          type: integer
          minimum: 0
        first_seen_at:
          type: string
          format: date-time
        actor_name:
          type: string
        malware_family:
          type: string
        external_id:
          type: string
          description: Idempotencia — ID del módulo Make
        metadata:
          type: object
          additionalProperties: true
    Institution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        sector:
          type: string
        country:
          type: string
        domain_obfuscated:
          type: string
    ExposureEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        summary:
          type: string
        severity:
          type: string
        status:
          type: string
        institution_slug:
          type: string
        institution_name:
          type: string
        actor_name:
          type: string
        credentials_count:
          type: integer
        first_seen_at:
          type: string
          format: date-time
    CitizenCheckResult:
      type: object
      properties:
        exposed:
          type: boolean
        events:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              title:
                type: string
              institution_name:
                type: string
        recommendations:
          type: array
          items:
            type: string
        mock:
          type: boolean
    Playbook:
      type: object
      properties:
        slug:
          type: string
        title_es:
          type: string
        body_md:
          type: string
        effort_hours:
          type: number
        cost_estimate_usd:
          type: number
        tags:
          type: array
          items:
            type: string
    TriageOutput:
      type: object
      required: [institution_slug, severity, suggested_title, suggested_summary, confidence, reasoning_brief]
      properties:
        institution_slug:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        suggested_title:
          type: string
        suggested_summary:
          type: string
        malware_family:
          type: string
          nullable: true
        credentials_count_estimate:
          type: integer
          minimum: 0
        confidence:
          type: number
          minimum: 0
          maximum: 1
        reasoning_brief:
          type: string
    TriageResponse:
      type: object
      properties:
        triage:
          $ref: '#/components/schemas/TriageOutput'
        mock:
          type: boolean
        run_id:
          type: string
        event_id:
          type: string
          format: uuid
        provider:
          type: string
    InvestigatorOutput:
      type: object
      required: [label, confidence, sources_summary, recommendation, reasoning_brief, hitl_required]
      properties:
        label:
          type: string
          enum: [confirmed, strong_evidence, claimed]
        confidence:
          type: number
        sources_summary:
          type: array
          items:
            type: object
            properties:
              source_type:
                type: string
              note:
                type: string
        recommendation:
          type: string
          enum: [approve_for_review, needs_more_info, reject]
        reasoning_brief:
          type: string
        hitl_required:
          type: boolean
    InvestigatorResponse:
      type: object
      properties:
        investigation:
          $ref: '#/components/schemas/InvestigatorOutput'
        mock:
          type: boolean
        run_id:
          type: string
        hitl_status:
          type: string
          enum: [pending_review]
        provider:
          type: string